PGranger2_BSALT.

PGranger2_BSALT (alternative bootstrap technique using two separate bootstrap DGPs for ystar) performs Wald test of Granger non-causality on VAR set up as SUR where the coefficients that are to be tested are the ones that do not show up in both Equations element.

PGranger2_BSALT(
  dat,
  Equations.full,
  Equations.H0,
  id.var,
  time.var,
  method = "SUR",
  reps = 1000,
  Tstar = 3,
  p,
  Model
)

Arguments

dat

A data frame.

Equations.full

The fully specified SUR model with a separate estimation equation for each country (incl. intercepts).

Equations.H0

The set of estimation equations under the null hypothesis. All variables not contained in Equations.H0 that are in Equations.full will be tested.

id.var

The cross-sectional identifier of the data set, e.g. the country code.

time.var

The time-series identifier of the data set, e.g. years or periods.

method

Currently only SUR supported.

reps

Number of bootstrap repetitions. Set to 1,000 by default.

Value

The routine performs either a equation-wise OLS or a two-step FGLS procedure respectively .

Examples

PGranger2_SUR(dat, Equations.full, Equations.H0, id.var, time.var, method = "SUR", reps = 10000)
#> Error in PGranger2_SUR(dat, Equations.full, Equations.H0, id.var, time.var,     method = "SUR", reps = 10000): unused argument (reps = 10000)